feat(core/rust): introduce trait PaginateFull throughout Delizia
PaginateFull uses Pager instead of reporting just the total number of
pages. Delizia will rely on this trait; going forward, we'll want
PaginateFull to replace Paginate, but this refactor would be too big if
we also needed to include Caesar and Bolt in it

+++ core/embed/rust/src/ui/layout_delizia/flow/util.rs
@@ -48,7 +48,7 @@ pub struct ConfirmValue {
     chunkify: bool,
     text_mono: bool,
     page_counter: bool,
-    page_limit: Option<usize>,
+    page_limit: Option<u16>,
     swipe_up: bool,
     swipe_down: bool,
     swipe_right: bool,
@@ -189,7 +189,7 @@ impl ConfirmValue {
         self
     }
 
-    pub const fn with_page_limit(mut self, page_limit: Option<usize>) -> Self {
+    pub const fn with_page_limit(mut self, page_limit: Option<u16>) -> Self {
         self.page_limit = page_limit;
         self
     }
